requirements.txt
matplotlib
reportlab
Pillow
pyinstaller